+++ src/bbochat/constants.py
@@ -13,13 +13,19 @@ APP_AUTHOR = "psionman"
 # HTML_DIR = resolve_path('html', __file__)
 HELP_URI = ""
 DATA_DIR = str(Path(user_data_dir(APP_NAME, AUTHOR)))
-
 # Config
 CONFIG_DIR = user_config_dir(APP_NAME, APP_AUTHOR)
 CONFIG_PATH = Path(CONFIG_DIR, "config.toml")
 USER_DATA_DIR = user_data_dir(APP_NAME, APP_AUTHOR)
 DOCS_DIR = userpaths.get_my_documents()
 STATE_DIR = user_state_dir(APP_NAME, APP_AUTHOR)
+PSIUTILS_DIR = user_data_dir("psiutils", APP_AUTHOR)
+
+# Buttons
+BUTTONS_DIR = Path(PSIUTILS_DIR, "buttons")
+BUTTON_ICON_PATH = str(Path(BUTTONS_DIR, "icons"))
+BUTTON_CONFIG_PATH = str(Path(BUTTONS_DIR, "buttons.json"))
+
 
 # GUI
 APP_TITLE = "BBO Chat"