+++ src/bbochat/forms/frm_notes.py
@@ -82,7 +82,7 @@ class NotesFrame:
     def _button_frame(self, master: ttk.Frame) -> tk.Frame:
         frame = ButtonFrame(master, tk.VERTICAL)
         frame.buttons = [
-            frame.icon_button("new", self._new),
+            frame.icon_button("new-note", self._new),
             frame.icon_button("edit", self._edit, True),
             frame.icon_button("delete", self._delete, True),
         ]