+++ src/bbochat/forms/frm_partners.py
@@ -132,7 +132,7 @@ class PartnerFrame:
     def _button_frame(self, master: ttk.Frame) -> tk.Frame:
         frame = ButtonFrame(master, tk.VERTICAL)
         frame.buttons = [
-            frame.icon_button("new", self._new),
+            frame.icon_button("new-user", self._new),
             frame.icon_button("edit", self._edit, True),
             frame.icon_button("delete", self._delete, True),
         ]