+++ HISTORY.md
@@ -14,6 +14,7 @@
 11. Don't add CHAT mode messages to history
 12. Rename message.py to message_store.py
 13. Introduce state.py for application state management
+14. Reconfigure buttons
 
 ## Version 0.0.8 01 September 2026
 1. Implement message.py as dynamic data repository